Fix hashmapRemove by introducing a DUMMY map entry
I initially failed to account for how linear probing works when writing
the logic for removing entries. It used to free the entry and replace it
with NULL and be done. This causes the linear probe to fail under
certain circumstances when performng a lookup since it stops after
finding an empty bucket.

To fix this, hash maps typically utilize a "dummy" entry to represent an
open bucket that used to contain data but doesn't anymore. As long as
the linear probe lookup only stops on empty entries and not dummy
entries, then it will correctly probe forward to find elements even
though there is a "hole" in the middle of its search path.