Removed most NULL checks for internal functions.
All of the "public" API functions already check for erronous input
anyways, and all of the internal functions are only used within these
"public" API functions so we can guarantee no bad parameters will be
passed to them.

Why not save a handful of clock cycles here and there for no cost?