Add README and make compilation less annoying
The README goes over basic usage of the library and a couple of the
really important functions necessary to use it; as well as how to
compile the library.

The makefile now creates the bin directory if it doesn't exist, and has
an extra rule "hashmap" that places the compiled library into the bin
directory and does not move it to the parent directory.