Simplify index calculations (again)
I've abstracted the hash index calculations yet again. All index
calculation logic is now contained in the static function
_findSlotEntries. Given an array of HashEntry pointers, the length of
the array, and the hash value to find an entry for, it returns the index
where the associated key should be stored (or currently is stored if it
is already present in the array).

The static function _findSlotMap is identical in functionality except it
uses a HashMap struct to simplify the parameters necessary to call the
function.

From this change, the static _hashmapGetIndex and _hashmapInsert
functions have been simplified appreciably.